fix: protect all shopping cart endpoints
Only owner can access shopping cart endpoints

+++ packages/shopping/src/controllers/shopping-cart.controller.ts
@@ -48,17 +48,22 @@ export class ShoppingCartController {
       },
     },
   })
+  @authenticate('jwt')
   async set(
     @param.path.string('userId') userId: string,
     @requestBody({description: 'shopping cart'}) cart: ShoppingCart,
   ): Promise<void> {
-    debug('Create shopping cart %s: %j', userId, cart);
-    if (userId !== cart.userId) {
-      throw new HttpErrors.BadRequest(
-        `User id does not match: ${userId} !== ${cart.userId}`,
-      );
+    if (this.currentUserProfile[securityId] === userId) {
+      debug('Create shopping cart %s: %j', userId, cart);
+      if (userId !== cart.userId) {
+        throw new HttpErrors.BadRequest(
+          `User id does not match: ${userId} !== ${cart.userId}`,
+        );
+      }
+      await this.shoppingCartRepository.set(userId, cart);
+    } else {
+      throw HttpErrors(401);
     }
-    await this.shoppingCartRepository.set(userId, cart);
   }
 
   /**
@@ -77,8 +82,8 @@ export class ShoppingCartController {
   async get(
     @param.path.string('userId') userId: string,
   ): Promise<ShoppingCart> {
-    debug('Get shopping cart %s', userId);
     if (this.currentUserProfile[securityId] === userId) {
+      debug('Get shopping cart %s', userId);
       const cart = await this.shoppingCartRepository.get(userId);
       debug('Shopping cart %s: %j', userId, cart);
       if (cart == null) {
@@ -104,9 +109,14 @@ export class ShoppingCartController {
       },
     },
   })
+  @authenticate('jwt')
   async remove(@param.path.string('userId') userId: string): Promise<void> {
-    debug('Remove shopping cart %s', userId);
-    await this.shoppingCartRepository.delete(userId);
+    if (this.currentUserProfile[securityId] === userId) {
+      debug('Remove shopping cart %s', userId);
+      await this.shoppingCartRepository.delete(userId);
+    } else {
+      throw HttpErrors(401);
+    }
   }
 
   /**
@@ -124,11 +134,16 @@ export class ShoppingCartController {
       },
     },
   })
+  @authenticate('jwt')
   async addItem(
     @param.path.string('userId') userId: string,
     @requestBody({description: 'shopping cart item'}) item: ShoppingCartItem,
   ): Promise<ShoppingCart> {
-    debug('Add item %j to shopping cart %s', item, userId);
-    return this.shoppingCartRepository.addItem(userId, item);
+    if (this.currentUserProfile[securityId] === userId) {
+      debug('Add item %j to shopping cart %s', item, userId);
+      return this.shoppingCartRepository.addItem(userId, item);
+    } else {
+      throw HttpErrors(401);
+    }
   }
 }
